requirements.txt
aiohttp==3.8.3
aioresponses==0.7.4
aiosignal==1.3.1
ajpy==0.0.4
alabaster==0.7.12
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Automat==20.2.0
Babel==2.10.3
backcall==0.2.0
backdoor-factory==0.0.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
beniget==0.4.1
blinker==1.4
bluepy==1.3.0
Brlapi==0.8.4
Brotli==1.0.9
cairocffi==1.4.0
capstone==4.0.2
catfish==4.16.4
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.6
chirp==0.3.0.dev0
click==8.0.3
click-plugins==1.1.1
colorama==0.4.5
constantly==15.1.0
contextlib2==0.6.0.post1
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.3.2
debtags==2.1
decorator==5.1.1
Deprecated==1.2.13
dicttoxml==1.7.4
distro==1.8.0
dnspython==2.2.1
docutils==0.17.1
exceptiongroup==1.1.0
executing==1.2.0
flasgger==0.9.5
Flask==2.0.3
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
fs==2.4.16
future==0.18.2
gast==0.5.2
GDAL==3.6.2
GeoIP==1.3.2
geoip2==2.9.0
gevent==21.12.0
greenlet==1.1.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpretty==1.1.4
hyperframe==6.0.0
hyperlink==21.0.0
idna==3.3
imageio==2.4.1
imagesize==1.4.1
impacket==0.10.0
importlib-metadata==4.12.0
incremental==21.3.0
iniconfig==2.0.0
invoke==1.7.0
ipdb==0.13.9
IPy==1.1
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.0.3
jsonpointer==2.3
jsonschema==4.7.2
kaitaistruct==0.9
kali-tweaks==2022.4.1
KismetCaptureBtGeiger==2021.7.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2020.10.1
KismetCaptureRtladsb==2020.10.1
KismetCaptureRtlamr==2020.10.1
kismetexternal==2019.12.1
kiwisolver==0.0.0
ldap3==2.9.1
ldapdomaindump==0.9.3
lightdm-gtk-greeter-settings==1.2.2
louis==3.23.0
lxml==4.9.1
lz4==4.0.2+dfsg
M2Crypto==0.38.0
Mako==1.2.2
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
maxminddb==2.1.0
mechanize==0.4.8
mistune0==0.8.4
mitmproxy==8.1.1
more-itertools==8.10.0
mpmath==0.0.0
msgpack==1.0.3
multidict==6.0.4
mysqlclient==1.4.6
netaddr==0.8.0
netifaces==0.11.0
networkx==2.8.8
ntpsec==1.2.1
numpy==1.23.5
olefile==0.46
onboard==1.4.1
outcome==1.2.0
packaging==21.3
paramiko==2.12.0
parso==0.8.3
passlib==1.7.4
patator==0.9
pcapy==0.11.5.dev0
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
plecost==1.1.2
pluggy==1.0.0
pluginbase==1.0.1
ply==3.11
prettytable==3.6.0
prompt-toolkit==3.0.36
protobuf==3.12.4
psycopg2==2.9.4
ptyprocess==0.7.0
publicsuffix2==2.20191221
pure-eval==0.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycurl==7.45.2
pydig==0.4.0
pydot==1.4.2
pygccxml==2.2.1
pygeoip==0.3.2
Pygments==2.12.0
PyGObject==3.42.2
pygraphviz==1.7
PyHamcrest==2.0.3
pyinotify==0.9.6
PyNaCl==1.5.0
pynvim==0.4.2
PyOpenGL==3.1.6
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyperclip==1.8.2
PyQt-Qwt==1.2.2
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQt6==6.4.0
PyQt6-sip==13.4.0
pyqtgraph==0.13.1
pyrsistent==0.18.1
pyrtlsdr==0.2.91
pyserial==3.5
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pytest==7.2.1
pytest-warnings==0.3.1
python-apt===2.3.0-nmu1-b1
python-dateutil==2.8.2
python-debian==0.1.48
python-dotenv==0.21.0
python-magic==0.4.26
python-Wappalyzer==0.3.1
python-whois==0.8.0
pythran==0.11.0
pytz==2022.6
pyusb==1.2.1.post1
pyxdg==0.27
PyYAML==6.0
pyzmq==24.0.1
redis==4.3.4
requests==2.27.1
requests-toolbelt==0.9.1
rfc3987==1.3.8
rfcat==1.9.7
roman==3.3
rq==1.11.1
rtlsdr-scanner==1.3.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.7
scapy==2.4.5+g9420c22
schema==0.7.5
SciPy==1.8.1
selenium==4.8.0
service-identity==18.1.0
shodan==1.28.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
specan==0.0.0
Sphinx==5.3.0
stack-data==0.6.2
sympy==1.11.1
termcolor==1.1.0
thrift==0.17.0
tld==0.11.11
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.4.0
ufoLib2==0.14.0
unicodecsv==0.14.1
unicodedata2==15.0.0
uritemplate==4.1.1
urllib3==1.26.12
urwid==2.1.2
visvis==1.13.0
wafw00f==2.2.0
wapiti3==3.0.4
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websockets==10.2
Werkzeug==2.0.2
wfuzz==3.1.0
whois==0.9.25
wifite==2.6.6
wrapt==1.14.1
wsproto==1.1.0
wxPython==4.2.0
xcffib==0.11.1
xdg==5
XlsxWriter==3.0.2
yarl==1.8.2
yaswfp==0.9.3
zipp==1.0.0
zope.event==4.4
zope.interface==5.4.0
